feat: Structuring testing code correctly Separating functions with commands in command.rs file, Utility functions in utils.rs, Putting all structs and traits in types.rs file.
